Fix autoload issues in subscriber
