+++ NEWS.md
@@ -75,6 +75,7 @@ sudo make install
 - Added dynamic meta title and canonical link to improve SEO in public pages for the Spatial Data Catalog [#16157](https://github.com/CartoDB/cartodb/pull/16157)
 - Fixed error that added multiple canonical links in the Spatial Data Catalog [#16160](https://github.com/CartoDB/cartodb/pull/16160)
 - Modify deprecation warning for DO analysis in builder [#16163](https://github.com/CartoDB/cartodb/pull/16163)
+- Fix autoload issues in subscriber [#16171](https://github.com/CartoDB/cartodb/pull/16171)
 
 4.44.0 (2020-11-20)
 -------------------
