+++ app/models/carto/api_key.rb
@@ -1,23 +1,9 @@
 require 'securerandom'
+require './lib/api_key_grants_validator'
 require_dependency 'carto/errors'
 require_dependency 'carto/helpers/auth_token_generator'
 require_dependency 'carto/oauth_provider/scopes/scopes'
 require_dependency 'carto/api_key_permissions'
-
-class ApiKeyGrantsValidator < ActiveModel::EachValidator
-  def validate_each(record, attribute, value)
-    return record.errors[attribute] = ['grants has to be an array'] unless value&.is_a?(Array)
-
-    record.errors[attribute] << 'only one apis section is allowed' unless value.count { |v| v[:type] == 'apis' } == 1
-
-    max_one_sections = ['database', 'dataservices', 'user', 'data-observatory']
-    max_one_sections.each do |section|
-      if value.count { |v| v[:type] == section } > 1
-        record.errors[attribute] << "only one #{section} section is allowed"
-      end
-    end
-  end
-end
 
 module Carto
   class ApiKey < ActiveRecord::Base
